Fixed archive naming

+++ WinImageBuilder.psm1
@@ -683,7 +683,7 @@ function Compress-Image {
     if (!(Test-Path $VirtualDiskPath)) {
         Throw "$VirtualDiskPath not found"
     }
-    $tmpName = $ImagePath + "." + (Get-Random)
+    $tmpName = '{0}.tar' -f @((Get-PathWithoutExtension($ImagePath)))
 
     $7zip = Get-7zipPath
     $pigz = Get-PigzPath
@@ -726,6 +726,9 @@ function Compress-Image {
         Remove-Item -Force $tmpName -ErrorAction SilentlyContinue
         Remove-Item -Force $VirtualDiskPath -ErrorAction SilentlyContinue
         throw
+    }
+    if (Test-Path $ImagePath) {
+        throw "File $ImagePath already exists. The image has been created at $tmpName."
     }
     Move-Item ($tmpName + ".gz") $ImagePath
     Write-Output "MaaS image is ready and available at: $ImagePath"
